Load player's st_channel from JSON on Player object creation.  Right now the field is optional and None is a valid value.
Eventually I plan on adding a command to the bot to allow an ST to either manually link a player to a channel, or to automatically create a new channel that is autmatically linked.

Once that's in place I plan on checking that all players in a game have a channel, and if not, automatically creating channels for them before allowing the game to start.